Update 1.6.0
Updates:

Added support for 2 Listeners:

ObjectChangeListener is an Interface which can be implemented in an object that is used trough the ReflectionGrid. The function this interface has gets called, when the object inside TransferGrid has a value changed.

ChangeListener is an Interface which can be implemented in an object that is NOT the object within TransferGrid. This could be used, for example, in combination with the GUI Manager to let it know which object had which field updated, and then only refresh key parts of the GUI.