Added 2 new Interfaces for 1.6.0
This update adds 2 new Interfaces for Version 1.6.0

ObjectChangeListener
This Interface gets implemented into the objects that are used trough TransferGrid.
This allows the TransferGrid to notify the object of which field has been updated.

ChangeListener
This Interface gets implemented in other objects. Example would be the GUI, which would get notified with the object and the field every time a reflective change happens.